Data caching responsibilities are moved into the DataCache class
This represents the second BimodalCache implementation we've introduced
These types and related types have been marked internal for now
[#1219]
